# demo-birthday-alert.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "demo-birthday-alert",
    "description": "",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "User"
    }
  ],
  "paths": {
    "/v1/user": {
      "post": {
        "summary": "Store",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": ["User"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apidog-ignore-properties": [],
                "x-apidog-orders": []
              },
              "example": {
                "email": "jhonny.cage@email.io",
                "name": "Jhonny Cage",
                "date_of_birth": "1999-02-22",
                "location": "Swansea"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/465993/apis/api-5676090-run",
        "security": []
      }
    },
    "/v1/user/{id}": {
      "delete": {
        "summary": "Delete",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": ["User"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/465993/apis/api-5676914-run",
        "security": []
      },
      "put": {
        "summary": "Update",
        "x-apidog-folder": "User",
        "x-apidog-status": "developing",
        "deprecated": false,
        "description": "",
        "tags": ["User"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "1",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {},
                "x-apidog-ignore-properties": [],
                "x-apidog-orders": []
              },
              "example": {
                "location": "Jayapura"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {},
                  "x-apidog-ignore-properties": [],
                  "x-apidog-orders": []
                }
              }
            }
          }
        },
        "x-run-in-apidog": "https://apidog.com/web/project/465993/apis/api-5717676-run",
        "security": []
      }
    }
  },
  "components": {
    "schemas": {},
    "securitySchemes": {}
  },
  "servers": [
    {
      "url": "http://demo-birthday-alert.lndo.site/api",
      "description": "demo-birthday-alert"
    }
  ]
}